move text direction (RTL, bidi) and segmentation (curved label segments, Arabic) to separate file
